made button that fixes deleted pieces and non-joining pices
This algorithm has a record of all valid GUIDs the puzzle started with.  If it finds any of the GUIDs are missing it respawns them.  Also, if a piece is not part of the initial valid GUIDs, then it will detach and delete it. However if the unknow GUID is a parent of many pieces (> 10), then keep it and make it a valid GUID.