fix getPiece() function
getPiece() has calls on lines 132, 152, 188, 216, 300, 584, 770, 787 and 1203
Some of which pass a TTS object as an argument and some, a pieceId as an argument.

getPiece() can handle an object not in "pieceCache" it caches it properly, but if it encounters a pieceId not in "pieceCache" then it errors and doesn't join a dropped piece.
The error is generated on line 950
"if object.tag == nil or object.tag ~= 'Generic' then return end"
of trying to index a number.

The fix is to handle both an object and pieceId separately and return a valid piece table for both(or nil if no such piece exists)